Write lock doesn't depend on pending read locks
The while loop polling for write lock aquiring only
matters because there can be multiple readers.  Because
there can only be a single writer, we don't need to
account for this when acquiring a write lock, the normal
locking mechanism takes care of this.

Verified the change with the scripts/threadedtest.py script.